api_server/requirements.txt
alembic==1.8.1; python_version >= "3.7"
anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
asyncpg==0.26.0; python_full_version >= "3.6.0"
bcrypt==4.0.0; python_version >= "3.6"
certifi==2022.9.24; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
databases==0.6.1; python_version >= "3.7"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
fastapi==0.85.0; python_version >= "3.7"
greenlet==1.1.3; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
h11==0.12.0; python_version >= "3.7"
httpcore==0.15.0; python_version >= "3.7"
httpx==0.23.0; python_version >= "3.7"
idna==3.4; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.2"
importlib-metadata==5.0.0; python_version < "3.9" and python_version >= "3.7"
importlib-resources==5.9.0; python_version < "3.9" and python_version >= "3.7"
mako==1.2.3; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
motor==3.0.0; python_version >= "3.7"
passlib==1.7.4
psycopg2==2.9.4; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pydantic==1.10.2; python_version >= "3.7"
pymongo==4.2.0; python_version >= "3.7"
python-dotenv==0.21.0; python_version >= "3.7"
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1; python_version >= "3.7" and python_version < "4"
rfc3986==1.5.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
sqlalchemy==1.4.41; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette==0.20.4; python_version >= "3.7"
typing-extensions==4.3.0; python_version < "3.10" and python_version >= "3.7"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
uvicorn==0.18.3; python_version >= "3.7"
zipp==3.8.1; python_version < "3.9" and python_version >= "3.7"
